demos/vao_demo: pass literal pointers through uintptr_t
This avoids a compiler-warning on MSVC.

+++ src/demos/vao_demo.c
@@ -286,8 +286,8 @@ static void Init( void )
    glEnableClientState( GL_COLOR_ARRAY );
 
    glBindVertexArrayAPPLE( 0 );
-   glVertexPointer( 4, GL_FLOAT, sizeof(GLfloat) * 4, (void *) 0xDEADBEEF );
-   glColorPointer( 4, GL_FLOAT, sizeof(GLfloat) * 4, (void *) 0xBADDC0DE );
+   glVertexPointer( 4, GL_FLOAT, sizeof(GLfloat) * 4, (void *) (uintptr_t) 0xDEADBEEF );
+   glColorPointer( 4, GL_FLOAT, sizeof(GLfloat) * 4, (void *) (uintptr_t) 0xBADDC0DE );
 }
